+++ vizro-ai/src/vizro_ai/dashboard/response_models/dashboard.py
@@ -7,7 +7,7 @@ try:
     from pydantic.v1 import BaseModel, Field
 except ImportError:  # pragma: no cov
     from pydantic import BaseModel, Field
-from vizro_ai.dashboard.plan.page import PagePlanner
+from vizro_ai.dashboard.response_models.page import PagePlanner
 
 logger = logging.getLogger(__name__)
